Rename dev-build to debug-build, remove update package

+++ build/build.php
@@ -41,7 +41,7 @@ function usage(string $command)
     echo PHP_TAB . PHP_TAB . '--include-bzip2:' . PHP_TAB . PHP_TAB . 'Exclude the generation of .tar.bz2 packages' . PHP_EOL;
     echo PHP_TAB . PHP_TAB . '--exclude-zstd:' . PHP_TAB . PHP_TAB . PHP_TAB . 'Include the generation of .tar.zst packages' . PHP_EOL;
     echo PHP_TAB . PHP_TAB . '--disable-patch-packages:' . PHP_TAB . 'Disable the generation of patch packages' . PHP_EOL;
-    echo PHP_TAB . PHP_TAB . '--dev-build:' . PHP_TAB . 'Include development packages and build folder' . PHP_EOL;
+    echo PHP_TAB . PHP_TAB . '--debug-build:' . PHP_TAB . 'Include development packages and build folder' . PHP_EOL;
     echo PHP_TAB . PHP_TAB . '--help:' . PHP_TAB . PHP_TAB . PHP_TAB . PHP_TAB . 'Show this help output' . PHP_EOL;
     echo PHP_EOL;
 }
@@ -237,10 +237,10 @@ $tmp      = $here . '/tmp';
 $fullpath = $tmp . '/' . $time;
 
 // Parse input options
-$options = getopt('', ['help', 'remote::', 'exclude-zip', 'exclude-gzip', 'include-bzip2', 'exclude-zstd', 'dev-build', 'disable-patch-packages']);
+$options = getopt('', ['help', 'remote::', 'exclude-zip', 'exclude-gzip', 'include-bzip2', 'exclude-zstd', 'debug-build', 'disable-patch-packages']);
 
 $remote             = $options['remote'] ?? false;
-$devBuild           = $options['dev-build'] ?? false;
+$debugBuild           = isset($options['debug-build']);
 $excludeZip         = isset($options['exclude-zip']);
 $excludeGzip        = isset($options['exclude-gzip']);
 $excludeBzip2       = !isset($options['include-bzip2']);
@@ -268,7 +268,7 @@ if (!$remote) {
 }
 
 $composerOptions = ' ';
-if (!$devBuild) {
+if (!$debugBuild) {
     $composerOptions .= '--no-dev';
 }
 
@@ -326,7 +326,7 @@ if ($verReturnCode !== 0) {
 }
 
 // Clean the checkout of extra resources
-if (!$devBuild) {
+if (!$debugBuild) {
     clean_checkout($fullpath);
 }
 
@@ -334,12 +334,12 @@ if (!$devBuild) {
 system('composer dump-autoload --optimize --no-scripts' . $composerOptions);
 
 // Clean the Composer manifests now
-if (!$devBuild) {
+if (!$debugBuild) {
     clean_composer($fullpath);
 }
 
 // And cleanup the Node installation
-if (!$devBuild) {
+if (!$debugBuild) {
     system('rm -rf node_modules');
 }
 
@@ -466,7 +466,7 @@ $packageStability = str_replace(' ', '_', Version::DEV_STATUS);
 // Delete the files and folders we exclude from the packages (tests, docs, build, etc.).
 echo "Delete folders not included in packages.\n";
 
-if (!$devBuild) {
+if (!$debugBuild) {
     foreach ($doNotPackage as $removeFile) {
         system('rm -rf ' . $time . '/' . $removeFile);
     }
@@ -515,7 +515,7 @@ for ($num = $release - 1; $num >= 0; $num--) {
         }
 
 
-        if (!$devBuild && ($dirtyHackForMediaCheck || $doNotPackageFile || $doNotPatchFile || $doNotPackageBaseFolder || $doNotPatchBaseFolder)) {
+        if (!$debugBuild && ($dirtyHackForMediaCheck || $doNotPackageFile || $doNotPatchFile || $doNotPackageBaseFolder || $doNotPatchBaseFolder)) {
             continue;
         }
 
@@ -636,45 +636,47 @@ if (!$excludeZstd) {
 }
 
 // Create full update file without the default logs directory, installation folder, or sample images.
-echo "Build full update package.\n";
-system('rm -r administrator/logs');
-system('rm -r installation');
-system('rm -r images/banners');
-system('rm -r images/headers');
-system('rm -r images/sampledata');
-system('rm images/joomla_black.png');
-system('rm images/powered_by.png');
+if (!$debugBuild) {
+    echo "Build full update package.\n";
+    system('rm -r administrator/logs');
+    system('rm -r installation');
+    system('rm -r images/banners');
+    system('rm -r images/headers');
+    system('rm -r images/sampledata');
+    system('rm images/joomla_black.png');
+    system('rm images/powered_by.png');
 
-if (!$excludeZip) {
-    $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.zip';
-    echo "Building " . $packageName . "... ";
-    system('zip -r ../packages/' . $packageName . ' * > /dev/null');
-    echo "done.\n";
-    $checksums[$packageName] = [];
-}
+    if (!$excludeZip) {
+        $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.zip';
+        echo "Building " . $packageName . "... ";
+        system('zip -r ../packages/' . $packageName . ' * > /dev/null');
+        echo "done.\n";
+        $checksums[$packageName] = [];
+    }
 
-if (!$excludeGzip) {
-    $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.tar.gz';
-    echo "Building " . $packageName . "... ";
-    system('tar --create --gzip --file ../packages/' . $packageName . ' * > /dev/null');
-    echo "done.\n";
-    $checksums[$packageName] = [];
-}
+    if (!$excludeGzip) {
+        $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.tar.gz';
+        echo "Building " . $packageName . "... ";
+        system('tar --create --gzip --file ../packages/' . $packageName . ' * > /dev/null');
+        echo "done.\n";
+        $checksums[$packageName] = [];
+    }
 
-if (!$excludeBzip2) {
-    $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.tar.bz2';
-    echo "Building " . $packageName . "... ";
-    system('tar --create --bzip2 --file ../packages/' . $packageName . ' * > /dev/null');
-    echo "done.\n";
-    $checksums[$packageName] = [];
-}
+    if (!$excludeBzip2) {
+        $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.tar.bz2';
+        echo "Building " . $packageName . "... ";
+        system('tar --create --bzip2 --file ../packages/' . $packageName . ' * > /dev/null');
+        echo "done.\n";
+        $checksums[$packageName] = [];
+    }
 
-if (!$excludeZstd) {
-    $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.tar.zst';
-    echo "Building " . $packageName . "... ";
-    system('tar "-I zstd --ultra -22" --create --file ../packages/' . $packageName . ' * > /dev/null');
-    echo "done.\n";
-    $checksums[$packageName] = [];
+    if (!$excludeZstd) {
+        $packageName = 'Joomla_' . $fullVersion . '-' . $packageStability . '-Update_Package.tar.zst';
+        echo "Building " . $packageName . "... ";
+        system('tar "-I zstd --ultra -22" --create --file ../packages/' . $packageName . ' * > /dev/null');
+        echo "done.\n";
+        $checksums[$packageName] = [];
+    }
 }
 
 chdir('..');
